+++ src/com/mideas/rpg/v2/command/CommandDeleteCharacter.java
@@ -14,7 +14,7 @@ public class CommandDeleteCharacter extends Command {
 	}
 	
 	public static void write(int id) {
-		ConnectionManager.getConnection().writeByte(DELETE_CHARACTER);
+		ConnectionManager.getConnection().writeShort(DELETE_CHARACTER);
 		ConnectionManager.getConnection().writeInt(id);
 		ConnectionManager.getConnection().send();
 	}
