+++ src/com/mideas/rpg/v2/command/CommandPing.java
@@ -12,13 +12,13 @@ public class CommandPing extends Command {
 	@Override
 	public void read() {
 		Mideas.setPing((System.currentTimeMillis()-timer));
-		ConnectionManager.getConnection().writeByte(PacketID.PING_CONFIRMED);
+		ConnectionManager.getConnection().writeShort(PacketID.PING_CONFIRMED);
 		ConnectionManager.getConnection().send();
 		pingSent = false;
 	}
 	
 	public static void write() {
-		ConnectionManager.getConnection().writeByte(PacketID.PING);
+		ConnectionManager.getConnection().writeShort(PacketID.PING);
 		ConnectionManager.getConnection().send();
 		timer = System.currentTimeMillis();
 		pingSent = true;
