+++ src/com/mideas/rpg/v2/command/CommandSelectScreenLoadCharacters.java
@@ -28,7 +28,7 @@ public class CommandSelectScreenLoadCharacters extends Command {
 	
 	public static void write() {
 		if(ConnectionManager.isConnected()) {
-			ConnectionManager.getConnection().writeByte(SELECT_SCREEN_LOAD_CHARACTERS);
+			ConnectionManager.getConnection().writeShort(SELECT_SCREEN_LOAD_CHARACTERS);
 			ConnectionManager.getConnection().send();
 		}
 	}
