+++ src/com/mideas/rpg/v2/command/item/CommandWeapon.java
@@ -15,7 +15,7 @@ public class CommandWeapon extends Command {
 
 	public static void write(int id) {
 		if(!ConnectionManager.getItemRequested().containsKey(id)) {
-			ConnectionManager.getConnection().writeByte(PacketID.WEAPON);
+			ConnectionManager.getConnection().writeShort(PacketID.WEAPON);
 			ConnectionManager.getConnection().writeInt(id);
 			ConnectionManager.getConnection().send();
 			ConnectionManager.getItemRequested().put(id, new Stuff(id));
